fixes #1215

+++ js/blocks.js
@@ -1187,6 +1187,10 @@ function Blocks (activity) {
         };
 
         var thisBlockobj = this.blockList[thisBlock];
+        if (thisBlockobj.name === 'vspace') {
+            return;
+        }
+
         if (thisBlockobj.name === 'rest2') {
             this._deletePitchBlocks(thisBlock);
         } else {
